// src/api/generated/model/createSkillRelationRequestRelationType.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type CreateSkillRelationRequestRelationType = typeof CreateSkillRelationRequestRelationType[keyof typeof CreateSkillRelationRequestRelationType];


export const CreateSkillRelationRequestRelationType = {
  SIMILAR_TO: 'SIMILAR_TO',
  PARENT_OF: 'PARENT_OF',
  PREREQUISITE_OF: 'PREREQUISITE_OF',
} as const;