// src/api/generated/model/skillRelationDto.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { SkillRelationDtoRelationType } from './skillRelationDtoRelationType';
import type { SkillRelationDtoSource } from './skillRelationDtoSource';

export interface SkillRelationDto {
  id?: string;
  fromSkillId?: string;
  fromSkillName?: string;
  toSkillId?: string;
  toSkillName?: string;
  relationType?: SkillRelationDtoRelationType;
  transferPenalty?: number;
  source?: SkillRelationDtoSource;
}